chore: reduce the scope of changes, avoid any breaking changes
