+++ src/__tests__/decodePointer.ts
@@ -7,4 +7,7 @@ test('decodePointer', () => {
   expect(decodePointer('paths/~1users')).toEqual('paths//users');
   expect(decodePointer('paths/foo~0users')).toEqual('paths/foo~users');
   expect(decodePointer('#/foo')).toEqual('#/foo');
+  expect(decodePointer('#/foo%20%5E%20bar')).toEqual('#/foo ^ bar');
+  expect(decodePointer('#/users% ')).toEqual('#/users% ');
+  expect(decodePointer('#/users%%20%5E%')).toEqual('#/users% ^%');
 });
