+++ src/__tests__/pathToPointer.ts
@@ -11,4 +11,7 @@ test('pathToPointer', () => {
   expect(pathToPointer(['paths', '/user/{userId}'])).toEqual('#/paths/~1user~1%7BuserId%7D');
   expect(pathToPointer(['paths', '/user/{userId}/~foo'])).toEqual('#/paths/~1user~1%7BuserId%7D~1~0foo');
   expect(pathToPointer(['$defs', 'User Model'])).toEqual('#/%24defs/User%20Model');
+  expect(pathToPointer(['\uD803\uDE6D', 'valid-pair'])).toEqual('#/%F0%90%B9%AD/valid-pair');
+  expect(pathToPointer(['\uD83D', 'unmatched-surrogate'])).toEqual('#/\uD83D/unmatched-surrogate');
+  expect(pathToPointer(['\uD83D', '\uDC00unmatched-surrogate'])).toEqual('#/\uD83D/\uDC00unmatched-surrogate');
 });
