+++ src/decodePointerFragment.ts
@@ -6,5 +6,5 @@ import { replaceInString } from './_utils';
  * decodePointer('#/paths/~1users) => '#/paths//users'
  */
 export const decodePointerFragment = (value: string): string => {
-  return decodeURIComponent(replaceInString(replaceInString(value, '~1', '/'), '~0', '~'));
+  return replaceInString(replaceInString(value, '~1', '/'), '~0', '~');
 };
