+++ src/encodePointer.ts
@@ -1,5 +1,4 @@
 import { replaceInString } from './_utils';
-import { encodeUriPointer } from './encodeUriPointer';
 
 /**
  * Sets special json pointer characters in a value. Example:
@@ -7,5 +6,5 @@ import { encodeUriPointer } from './encodeUriPointer';
  * encodePointer('#/paths//users) => '#/paths/~1users'
  */
 export const encodePointer = (value: string): string => {
-  return encodeUriPointer(replaceInString(replaceInString(value, '~', '~0'), '//', '/~1'));
+  return replaceInString(replaceInString(value, '~', '~0'), '//', '/~1');
 };
