+++ src/encodePointerUriFragment.ts
@@ -1,11 +1,15 @@
-import { replaceInString } from './_utils';
+import { Segment } from '@stoplight/types';
+
+import { encodePointerFragment } from './encodePointerFragment';
 import { encodeUriPointer } from './encodeUriPointer';
 
 /**
- * Escapes special json pointer characters in a value. Example:
+ * Escapes special json pointer characters in a value.
+ * Percent-encode characters.
  *
- * encodePointer('/paths/~users) => '~1paths~1~0users'
+ * @example encodePointer('/paths/~users) => '~1paths~1~0users'
  */
-export const encodePointerUriFragment = (value: string): string => {
-  return encodeUriPointer(replaceInString(replaceInString(value, '~', '~0'), '/', '~1'));
+export const encodePointerUriFragment = (value: Segment): Segment => {
+  const encoded = encodePointerFragment(value);
+  return typeof encoded === 'number' ? encoded : encodeUriPointer(encoded);
 };
